Updated workflow and setup script
The GitHub Actions workflow has been updated to include the checkout and installation of kcov. The dependencies installed by the workflow have also been expanded to include build-essential and cmake. The setup-action.sh script now includes steps for compiling and installing kcov, in addition to grabbing shellspec. The PATH is updated to include both shellspec and kcov.