Added script version display and improved shellspec installation
In this update, we've added a line to display the running script's name and version. This will provide more transparency about which version of the script is currently in use. Additionally, we've made a minor adjustment to the shellspec installation process by adding quotes around the path variable. This change ensures that paths with spaces are handled correctly, improving overall robustness of the installation process.