Added shell-utils plugin
A new bash script file, shell-utils.plugin.sh, has been added to the plugins directory. This script contains various bash functions and utilities for the Bloomshell project. It includes a check to prevent double sourcing of the script. The core library should be loaded before any additional scripts are loaded in this file.